:+1: Improve airtable_client
