+++ .vscode/settings.json
@@ -9,6 +9,7 @@
     "lako",
     "pytest",
     "q-lako",
+    "qlako",
     "yubi",
     "yubi key"
   ],
