+++ __init__.py
@@ -1,12 +1,18 @@
+from configparser import ConfigParser
+from datetime import datetime, timezone, timedelta
 import os
 
 from amazon.paapi import AmazonAPI
 from dotenv import load_dotenv
 from flask import Flask
 
+config = ConfigParser()
 load_dotenv(verbose=True)
 
 app = Flask(__name__)
+
+time_now = datetime.now(timezone.utc).astimezone(timezone(timedelta(hours=+9))).isoformat()
+
 
 amazon_api_client = AmazonAPI(os.getenv("amazon_access_key"),
                               os.getenv("amazon_secret_key"),
